IGNITE-10108: Refactored a test to avoid passing anonymous classes on compute. Put the method back to GridCommonAbstractTest.
